# services/control-plane/src/main/java/com/orvexa/controlplane/exception/ApiErrorResponse.java
package com.orvexa.controlplane.exception;

import java.time.OffsetDateTime;

public record ApiErrorResponse(
        OffsetDateTime timestamp,
        int status,
        String code,
        String message,
        String path,
        String correlationId
) {
}